tests: Move storage config to libstorage
