+++ tests/libstorage/BUILD.bazel
@@ -1,0 +1,9 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["config.go"],
+    importpath = "kubevirt.io/kubevirt/tests/libstorage",
+    visibility = ["//visibility:public"],
+    deps = ["//tests/flags:go_default_library"],
+)
