+++ tests/storage/BUILD.bazel
@@ -36,6 +36,7 @@ go_library(
         "//tests/framework/storage:go_default_library",
         "//tests/libnet:go_default_library",
         "//tests/libnode:go_default_library",
+        "//tests/libstorage:go_default_library",
         "//tests/libvmi:go_default_library",
         "//tests/util:go_default_library",
         "//vendor/github.com/google/goexpect:go_default_library",
